+++ src/browsers.ts
@@ -1,52 +1,99 @@
+import { exec } from "child_process"
+import { access, mkdir } from "fs/promises"
+import { homedir } from "os"
 import puppeteer, { Browser } from "puppeteer"
+import { resolve } from "path"
 
-let chromiumBrowser: Browser | null = null
-let firefoxBrowser: Browser | null = null
+let chrome: Browser | null = null
+let firefox: Browser | null = null
+let launchInProgress: Promise<void> = Promise.resolve ( )
+let installInProgress: Promise<void> = Promise.resolve ( )
 
-let launchPromise: Promise<void> | null = null  // Keep track of the launch promise
-
-// Function to launch both browsers asynchronously
-async function launchBrowsers ( ): Promise<void> {
-  if ( launchPromise ) {
-    return launchPromise
+async function installBrowsers ( ): Promise<void> {
+  try {
+    await access ( homedir ( ) + "/.cache" )
+  } catch {
+    await mkdir ( homedir ( ) + "/.cache" )
   }
 
-  launchPromise = ( async ( ) => {
-    try {
-      // Launch Chromium browser
-      chromiumBrowser = await puppeteer.launch ( {
-        headless: true, // Set to false if you want to see the browser
-        args: [ "--no-sandbox", "--disable-setuid-sandbox" ],  // Useful in certain environments (e.g., Docker)
-      } )
+  try {
+    await access ( homedir ( ) + "/.cache/puppeteer" )
+  } catch {
+    await mkdir ( homedir ( ) + "/.cache/puppeteer" )
+  }
 
-      // Launch Firefox browser
-      firefoxBrowser = await puppeteer.launch ( {
-        browser: "firefox", // Use Firefox
-        headless: true, // Set to false if you want to see the browser
-        args: [ "--no-sandbox", "--disable-setuid-sandbox" ],  // Useful in certain environments (e.g., Docker)
-      } )
-    } catch ( err ) {
-      console.error ( "Error launching browsers in @iqx-limited/quick-form:", err )
-    }
-  } ) ( )
-
-  return launchPromise
+  installInProgress = new Promise<void> ( async ( _resolve, reject ) => {
+    exec (
+      "npx --yes @puppeteer/browsers install chrome@stable && npx --yes @puppeteer/browsers install firefox@stable",
+      {
+        cwd: resolve ( homedir ( ), ".cache", "puppeteer" )
+      },
+      ( err, stdout ) => {
+        if ( err ) {
+          reject ( err )
+        } else {
+          console.log ( stdout )
+          _resolve ( )
+        }
+      }
+    )
+  } )
 }
 
-// Function to close both browsers
+async function getChromium ( ): Promise<Browser | null> {
+  await installInProgress
+  if ( chrome ) return chrome
+
+  try {
+    chrome = await puppeteer.launch ( {
+      args: [ "--no-sandbox", "--disable-setuid-sandbox" ],  // Useful in certain environments (e.g., Docker)
+    } )
+  } catch ( err ) {
+    console.error ( "Error launching Chromium browser in @iqx-limited/quick-form:", err )
+  }
+
+  return chrome
+}
+
+async function getFirefox ( ): Promise<Browser | null> {
+  await installInProgress
+  if ( firefox ) return firefox
+
+  try {
+    firefox = await puppeteer.launch ( {
+      browser: "firefox", // Use Firefox
+      headless: true, // Set to false if you want to see the browser
+      args: [ "--no-sandbox", "--disable-setuid-sandbox" ],  // Useful in certain environments (e.g., Docker)
+    } )
+  } catch ( err ) {
+    console.error ( "Error launching Firefox browser in @iqx-limited/quick-form:", err )
+  }
+
+  return firefox
+}
+
 async function closeBrowsers ( ): Promise<void> {
   try {
-    if ( chromiumBrowser ) {
-      await chromiumBrowser.close ( )
-      console.log ( "Chromium Browser Closed" )
+    if ( chrome ) {
+      await ( await chrome ).close ( )
+      chrome = null
     }
-    if ( firefoxBrowser ) {
-      await firefoxBrowser.close ( )
-      console.log ( "Firefox Browser Closed" )
+    if ( firefox ) {
+      await ( await firefox ).close ( )
+      firefox = null
     }
   } catch ( err ) {
     console.error ( "Error closing browsers in @iqx-limited/quick-form:", err )
   }
+}
+
+async function launchBrowsers ( ): Promise<void> {
+  launchInProgress = new Promise<void> ( async ( resolve ) => {
+    await installInProgress
+    await getChromium ( )
+    await getFirefox ( )
+    resolve ( )
+  } )
 }
 
 // Listen for process exit and close browsers
@@ -65,5 +112,10 @@ process.on ( "SIGTERM", async ( ) => {
   await closeBrowsers ( )
 } )
 
+async function checkForLaunching ( ) {
+  await installInProgress
+  await launchInProgress
+}
+
 // Export the browsers for use in other files
-export { launchBrowsers, chromiumBrowser, firefoxBrowser };
+export { getChromium, getFirefox, closeBrowsers, launchBrowsers, installBrowsers, checkForLaunching }