+++ src/index.ts
@@ -1,7 +1,7 @@
 import { satisfies } from "semver"
-import { launchBrowsers } from "./browsers.js"
+import { launchBrowsers, installBrowsers } from "./browsers.js"
 
-const requiredVersion = ">=22.0.0"
+const requiredVersion = ">=20.0.0"
 
 if ( !satisfies ( process.version, requiredVersion ) ) {
   console.error (
@@ -10,10 +10,14 @@ if ( !satisfies ( process.version, requiredVersion ) ) {
   process.exit ( 1 )
 }
 
-await launchBrowsers ( )
+installBrowsers ( ).then ( launchBrowsers ).catch ( e => {
+  console.error ( "Puppeteer installation failed in @iqx-limited/quick-form:", e )
+} )
 
 export * from "./modules/pdf2img.js"
 
 export * from "./modules/img2pdf.js"
 
 export * from "./modules/html2pdf.js"
+
+export { closeBrowsers, launchBrowsers } from "./browsers.js"